fix floating point exception (thanks igor)
The logarithm of a negative number is undefined. To store the
file length the int data type is not sufficient as the file
length is represented by a vlong (see /sys/include/libc.h:657).

Overflowing an integer causes it to be interpreted as a negative
number, that in turn causes an exception when trying to compute
the log of a negative number.

To fix the issue the type of the variable storing the largest
file length needs to change from int to vlong.